# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Smart API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Subject {
  id: number;
  name: string;
  color: string;
  icon: string;
  createdAt: string;
}

export interface CreateSubjectBody {
  name: string;
  color: string;
  icon: string;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  subjectId: number;
  subjectName?: string;
  lastUsedAt?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  title: string;
  content: string;
  subjectId: number;
}

export interface UpdateNoteBody {
  title?: string;
  content?: string;
  subjectId?: number;
}

export type GenerateQuizBodyDifficulty =
  (typeof GenerateQuizBodyDifficulty)[keyof typeof GenerateQuizBodyDifficulty];

export const GenerateQuizBodyDifficulty = {
  easy: "easy",
  normal: "normal",
  difficult: "difficult",
} as const;

export interface GenerateQuizBody {
  difficulty: GenerateQuizBodyDifficulty;
  /**
   * @minimum 3
   * @maximum 20
   */
  questionCount?: number;
}

export interface QuizQuestion {
  id: number;
  question: string;
  options: string[];
  correctAnswer: number;
  explanation: string;
}

export interface Quiz {
  noteId: number;
  difficulty: string;
  questions: QuizQuestion[];
}

export interface Schedule {
  id: number;
  subject: string;
  /**
   * @minimum 0
   * @maximum 6
   */
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  color: string;
  notificationEnabled: boolean;
  createdAt: string;
}

export interface CreateScheduleBody {
  subject: string;
  /**
   * @minimum 0
   * @maximum 6
   */
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  color: string;
  notificationEnabled: boolean;
}

export interface Goal {
  id: number;
  title: string;
  description?: string;
  deadline: string;
  completed: boolean;
  createdAt: string;
}

export interface CreateGoalBody {
  title: string;
  description?: string;
  deadline: string;
}

export interface UpdateGoalBody {
  title?: string;
  description?: string;
  deadline?: string;
  completed?: boolean;
}

export type MoodMood = (typeof MoodMood)[keyof typeof MoodMood];

export const MoodMood = {
  great: "great",
  good: "good",
  okay: "okay",
  tired: "tired",
  stressed: "stressed",
} as const;

export interface Mood {
  id: number;
  mood: MoodMood;
  note?: string;
  createdAt: string;
}

export type CreateMoodBodyMood =
  (typeof CreateMoodBodyMood)[keyof typeof CreateMoodBodyMood];

export const CreateMoodBodyMood = {
  great: "great",
  good: "good",
  okay: "okay",
  tired: "tired",
  stressed: "stressed",
} as const;

export interface CreateMoodBody {
  mood: CreateMoodBodyMood;
  note?: string;
}

export type ListNotesParams = {
  subjectId?: number;
};