# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Smart API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all subjects
 */
export const ListSubjectsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  color: zod.string(),
  icon: zod.string(),
  createdAt: zod.date(),
});
export const ListSubjectsResponse = zod.array(ListSubjectsResponseItem);

/**
 * @summary Create a new subject
 */
export const CreateSubjectBody = zod.object({
  name: zod.string(),
  color: zod.string(),
  icon: zod.string(),
});

/**
 * @summary Update a subject
 */
export const UpdateSubjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSubjectBody = zod.object({
  name: zod.string(),
  color: zod.string(),
  icon: zod.string(),
});

export const UpdateSubjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  color: zod.string(),
  icon: zod.string(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a subject
 */
export const DeleteSubjectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all notes
 */
export const ListNotesQueryParams = zod.object({
  subjectId: zod.coerce.number().optional(),
});

export const ListNotesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  subjectId: zod.number(),
  subjectName: zod.string().optional(),
  lastUsedAt: zod.date().optional(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListNotesResponse = zod.array(ListNotesResponseItem);

/**
 * @summary Create a new note
 */
export const CreateNoteBody = zod.object({
  title: zod.string(),
  content: zod.string(),
  subjectId: zod.number(),
});

/**
 * @summary Get a note by id
 */
export const GetNoteParams = zod.object({
  id: zod.coerce.number(),
});

export const GetNoteResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  subjectId: zod.number(),
  subjectName: zod.string().optional(),
  lastUsedAt: zod.date().optional(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a note
 */
export const UpdateNoteParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateNoteBody = zod.object({
  title: zod.string().optional(),
  content: zod.string().optional(),
  subjectId: zod.number().optional(),
});

export const UpdateNoteResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  subjectId: zod.number(),
  subjectName: zod.string().optional(),
  lastUsedAt: zod.date().optional(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a note
 */
export const DeleteNoteParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Generate a quiz for a note using AI
 */
export const GenerateQuizParams = zod.object({
  id: zod.coerce.number(),
});

export const generateQuizBodyQuestionCountMin = 3;
export const generateQuizBodyQuestionCountMax = 20;

export const GenerateQuizBody = zod.object({
  difficulty: zod.enum(["easy", "normal", "difficult"]),
  questionCount: zod
    .number()
    .min(generateQuizBodyQuestionCountMin)
    .max(generateQuizBodyQuestionCountMax)
    .optional(),
});

export const GenerateQuizResponse = zod.object({
  noteId: zod.number(),
  difficulty: zod.string(),
  questions: zod.array(
    zod.object({
      id: zod.number(),
      question: zod.string(),
      options: zod.array(zod.string()),
      correctAnswer: zod.number(),
      explanation: zod.string(),
    }),
  ),
});

/**
 * @summary List all study schedules
 */
export const listSchedulesResponseDayOfWeekMin = 0;
export const listSchedulesResponseDayOfWeekMax = 6;

export const ListSchedulesResponseItem = zod.object({
  id: zod.number(),
  subject: zod.string(),
  dayOfWeek: zod
    .number()
    .min(listSchedulesResponseDayOfWeekMin)
    .max(listSchedulesResponseDayOfWeekMax),
  startTime: zod.string(),
  endTime: zod.string(),
  color: zod.string(),
  notificationEnabled: zod.boolean(),
  createdAt: zod.date(),
});
export const ListSchedulesResponse = zod.array(ListSchedulesResponseItem);

/**
 * @summary Create a study schedule
 */
export const createScheduleBodyDayOfWeekMin = 0;
export const createScheduleBodyDayOfWeekMax = 6;

export const CreateScheduleBody = zod.object({
  subject: zod.string(),
  dayOfWeek: zod
    .number()
    .min(createScheduleBodyDayOfWeekMin)
    .max(createScheduleBodyDayOfWeekMax),
  startTime: zod.string(),
  endTime: zod.string(),
  color: zod.string(),
  notificationEnabled: zod.boolean(),
});

/**
 * @summary Update a schedule
 */
export const UpdateScheduleParams = zod.object({
  id: zod.coerce.number(),
});

export const updateScheduleBodyDayOfWeekMin = 0;
export const updateScheduleBodyDayOfWeekMax = 6;

export const UpdateScheduleBody = zod.object({
  subject: zod.string(),
  dayOfWeek: zod
    .number()
    .min(updateScheduleBodyDayOfWeekMin)
    .max(updateScheduleBodyDayOfWeekMax),
  startTime: zod.string(),
  endTime: zod.string(),
  color: zod.string(),
  notificationEnabled: zod.boolean(),
});

export const updateScheduleResponseDayOfWeekMin = 0;
export const updateScheduleResponseDayOfWeekMax = 6;

export const UpdateScheduleResponse = zod.object({
  id: zod.number(),
  subject: zod.string(),
  dayOfWeek: zod
    .number()
    .min(updateScheduleResponseDayOfWeekMin)
    .max(updateScheduleResponseDayOfWeekMax),
  startTime: zod.string(),
  endTime: zod.string(),
  color: zod.string(),
  notificationEnabled: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a schedule
 */
export const DeleteScheduleParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all goals
 */
export const ListGoalsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  deadline: zod.date(),
  completed: zod.boolean(),
  createdAt: zod.date(),
});
export const ListGoalsResponse = zod.array(ListGoalsResponseItem);

/**
 * @summary Create a goal
 */
export const CreateGoalBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
  deadline: zod.date(),
});

/**
 * @summary Update a goal
 */
export const UpdateGoalParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateGoalBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().optional(),
  deadline: zod.date().optional(),
  completed: zod.boolean().optional(),
});

export const UpdateGoalResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  deadline: zod.date(),
  completed: zod.boolean(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a goal
 */
export const DeleteGoalParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List mood check-ins
 */
export const ListMoodsResponseItem = zod.object({
  id: zod.number(),
  mood: zod.enum(["great", "good", "okay", "tired", "stressed"]),
  note: zod.string().optional(),
  createdAt: zod.date(),
});
export const ListMoodsResponse = zod.array(ListMoodsResponseItem);

/**
 * @summary Log a mood check-in
 */
export const CreateMoodBody = zod.object({
  mood: zod.enum(["great", "good", "okay", "tired", "stressed"]),
  note: zod.string().optional(),
});