# lib/api-zod/src/generated/types/createScheduleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Study Smart API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateScheduleBody {
  subject: string;
  /**
   * @minimum 0
   * @maximum 6
   */
  dayOfWeek: number;
  startTime: string;
  endTime: string;
  color: string;
  notificationEnabled: boolean;
}